+++ command/ssh/ssh.go
@@ -33,6 +33,14 @@ func NewCommand(home string, docker client.CommonAPIClient, output terminal.Outp
 		Short:   "SSH into a container",
 		Example: exampleText,
 		Args:    cobra.NoArgs,
+		PreRunE: func(cmd *cobra.Command, args []string) error {
+			// is the docker api alive?
+			if _, err := docker.Ping(cmd.Context()); err != nil {
+				return fmt.Errorf("Couldn’t connect to Docker; please make sure Docker is running.")
+			}
+
+			return nil
+		},
 		RunE: func(cmd *cobra.Command, args []string) error {
 			// get the current working directory
 			wd, err := os.Getwd()